fix windows trouble

+++ tests/test_pytest_arraydiff.py
@@ -17,10 +17,13 @@ def test_succeeds_func_default():
 def test_succeeds_func_text():
     return np.arange(3 * 5).reshape((3, 5))
 
+
 @pytest.mark.array_compare(file_format='pd_hdf', reference_dir=reference_dir)
 def test_succeeds_func_pdhdf():
     pd = pytest.importorskip('pandas')
-    return pd.DataFrame(data=np.arange(20), columns=['test_data'])
+    return pd.DataFrame(data=np.arange(20, dtype='int64'),
+                        columns=['test_data'])
+
 
 @pytest.mark.array_compare(file_format='fits', reference_dir=reference_dir)
 def test_succeeds_func_fits():
